style: admin view mobile (#58)
* Added mobile stylings for calendar view

* Added mobile stylings for pop up

* Add next day buttons

* pre PR comments

* rename files
